[AIP-21] Fix sftp_hook moving
