+++ airflow/contrib/operators/sftp_operator.py
@@ -17,16 +17,16 @@
 # specific language governing permissions and limitations
 # under the License.
 """
-This module is deprecated. Please use `airflow.providers.sftp.operators.sftp_operator`.
+This module is deprecated. Please use `airflow.providers.sftp.operators.sftp`.
 """
 
 import warnings
 
 # pylint: disable=unused-import
-from airflow.providers.sftp.operators.sftp_operator import SFTPOperator  # noqa
+from airflow.providers.sftp.operators.sftp import SFTPOperator  # noqa
 
 warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.sftp.operators.sftp_operator`.",
+    "This module is deprecated. Please use `airflow.providers.sftp.operators.sftp`.",
     DeprecationWarning,
     stacklevel=2,
 )
