+++ airflow/providers/sftp/sensors/sftp.py
@@ -21,7 +21,7 @@ This module contains SFTP sensor.
 """
 from paramiko import SFTP_NO_SUCH_FILE
 
-from airflow.providers.sftp.hooks.sftp_hook import SFTPHook
+from airflow.providers.sftp.hooks.sftp import SFTPHook
 from airflow.sensors.base_sensor_operator import BaseSensorOperator
 from airflow.utils.decorators import apply_defaults
